vendor CRUD check

+++ database/migrations/2022_11_29_003326_create_vendors_table.php
@@ -19,9 +19,13 @@ class CreateVendorsTable extends Migration
             $table->text('description');
             $table->string('country');
             $table->string('state');
-            $table->float('longitude');
-            $table->float('latitude');
+           // $table->float('longitude');
+           // $table->float('latitude');
+            $table->decimal('longitude', 10, 7);
+            $table->decimal('latitude', 10, 7);
             $table->timestamps();
+            
+
         });
     }
